requirements.txt
decouple
asgiref==3.2.10
boto3==1.14.51
botocore==1.17.51
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.4.0
django-heroku==0.3.1
django-storages==1.9.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
docutils==0.15.2
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
Pillow==7.2.0
psycopg2==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
stripe==2.51.0
urllib3==1.25.10
whitenoise==5.2.0
